basic-visualization/requirements.txt
astroid==4.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
flake8==7.3.0
fonttools==4.60.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mne==1.10.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
pyaml==25.7.0
pycodestyle==2.14.0
pyflakes==3.4.0
pylint==4.0.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
six==1.17.0
tomlkit==0.13.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
